:recycle: refactor: Merger jsx and tsx parsers into base version
